+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Fix
 
-- HTTP API: Extract API results dataframes separately and merge them after.
-- HTTP API: Add `data_length_filter` offset pagination config field to determine which part of data must be used to compute the data length.
+- HTTP API: API results are now correctly merged even if they need to be filtered or flattened. 
+
+### Added
+
+- HTTP API: Add `data_filter` offset pagination config field to determine which part of data must be used to compute the data length.
 
 ## [7.1.1] 2024-10-28